refactor API
