+++ api/views/documentation/event_info.yml
