+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
     container_name: fdo_dev_container
     ports:
       - "5000:5000"
-    command: ["python", "app.py"]
+    command: ["python", "-m", "api.app"]
     volumes:
       - .:/app
     env_file:
